Refactor to strategy.
