+++ src/test/java/ch/prinzbach/jass/validation/JassCardValidatorTest.java
@@ -77,6 +77,4 @@ public class JassCardValidatorTest {
         assertTrue(isCardValid);
     }
 
-
-
 }
